ref: move inputGroup to components/core/input
