+++ static/app/components/forms/fields/inputField.tsx
@@ -1,9 +1,9 @@
+import type {InputProps} from 'sentry/components/core/input/inputGroup';
+import {InputGroup} from 'sentry/components/core/input/inputGroup';
 import type {FormFieldProps} from 'sentry/components/forms/formField';
 import FormField from 'sentry/components/forms/formField';
 import FormFieldControlState from 'sentry/components/forms/formField/controlState';
 import type FormModel from 'sentry/components/forms/model';
-import type {InputProps} from 'sentry/components/inputGroup';
-import {InputGroup} from 'sentry/components/inputGroup';
 
 export interface InputFieldProps
   extends Omit<FormFieldProps, 'children'>,
