+++ static/app/components/forms/fields/textareaField.tsx
@@ -1,10 +1,10 @@
 import omit from 'lodash/omit';
 
+import type {TextAreaProps} from 'sentry/components/core/input/inputGroup';
+import {InputGroup} from 'sentry/components/core/input/inputGroup';
 import FormField from 'sentry/components/forms/formField';
 import FormFieldControlState from 'sentry/components/forms/formField/controlState';
 import type FormModel from 'sentry/components/forms/model';
-import type {TextAreaProps} from 'sentry/components/inputGroup';
-import {InputGroup} from 'sentry/components/inputGroup';
 
 // XXX(epurkhiser): This is wrong, it should not be inheriting these props
 import type {InputFieldProps} from './inputField';
