+++ static/app/components/numberInput.tsx
@@ -8,7 +8,7 @@ import {useNumberFieldState} from '@react-stately/numberfield';
 
 import {Button} from 'sentry/components/button';
 import type {InputStylesProps} from 'sentry/components/core/input';
-import {InputGroup} from 'sentry/components/inputGroup';
+import {InputGroup} from 'sentry/components/core/input/inputGroup';
 import {IconChevron} from 'sentry/icons/iconChevron';
 import {space} from 'sentry/styles/space';
 import mergeRefs from 'sentry/utils/mergeRefs';
