+++ static/app/views/settings/dynamicSampling/percentInput.tsx
@@ -3,7 +3,7 @@ import {forwardRef} from 'react';
 import {css} from '@emotion/react';
 import styled from '@emotion/styled';
 
-import {InputGroup} from 'sentry/components/inputGroup';
+import {InputGroup} from 'sentry/components/core/input/inputGroup';
 import {space} from 'sentry/styles/space';
 
 interface Props extends React.ComponentProps<typeof InputGroup.Input> {}
